+++ include/mimir/search/algorithms/siw.hpp
@@ -28,11 +28,11 @@ class SerializedIterativeWidthAlgorithm : public IAlgorithm
 {
 public:
     /// @brief Simplest construction
-    SerializedIterativeWidthAlgorithm(std::shared_ptr<IApplicableActionGenerator> applicable_action_generator, int max_arity);
+    SerializedIterativeWidthAlgorithm(std::shared_ptr<IApplicableActionGenerator> applicable_action_generator, size_t max_arity);
 
     /// @brief Complete construction
     SerializedIterativeWidthAlgorithm(std::shared_ptr<IApplicableActionGenerator> applicable_action_generator,
-                                      int max_arity,
+                                      size_t max_arity,
                                       std::shared_ptr<StateRepository> successor_state_generator,
                                       std::shared_ptr<IBrFSAlgorithmEventHandler> brfs_event_handler,
                                       std::shared_ptr<IIWAlgorithmEventHandler> iw_event_handler,
@@ -48,7 +48,7 @@ public:
 
 private:
     std::shared_ptr<IApplicableActionGenerator> m_aag;
-    int m_max_arity;
+    size_t m_max_arity;
 
     std::shared_ptr<StateRepository> m_ssg;
     std::shared_ptr<IBrFSAlgorithmEventHandler> m_brfs_event_handler;
